style: Split the files  Add Files  - Calculator.swift  - ValueError.swift  - String+Extension.swift
