+++ RacingCar/ValueError.swift
@@ -1,0 +1,14 @@
+//
+//  ValueError.swift
+//  RacingCar
+//
+//  Created by Mephrine on 2021/10/31.
+//
+
+import Foundation
+
+enum ValueError: Error {
+	case empty
+	case invalid
+	case invalidOperator
+}
